# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Post:
    properties:
      content:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
...
  model.User:
    properties:
      address:
        type: string
      age:
        type: string
      avatar:
        type: string
      experience:
        type: string
      gender:
        type: string
      id:
        type: integer
      identity:
        type: string
      intro:
        type: string
      nickName:
        type: string
      stature:
        type: string
    type: object
  utils.Error:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
  utils.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: 43.138.61.49
info:
  contact:
    email: 3264085417@qq.com
    name: BIG_DUST
  description: NursingAPI
  termsOfService: http://swagger.io/terrms/
  title: Nursing
  version: 1.1.0
paths:
...
  /api/v1/login:
    post:
      consumes:
      - multipart/form-data
      description: 要一个code
      parameters:
      - description: code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取用户信息
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/utils.Error'
      summary: 登录
      tags:
      - 用户管理
...
  /api/v1/user/{id}:
    get:
      consumes:
      - multipart/form-data
      description: 获取用户基本信息
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取用户信息
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/utils.Error'
      summary: 获取用户基本信息
      tags:
      - 用户管理
  /api/v1/user/{id}/collection:
    get:
      consumes:
      - multipart/form-data
      description: 获取收藏的帖子
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取收藏的帖子
          schema:
            items:
              $ref: '#/definitions/model.Post'
            type: array
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/utils.Error'
      summary: 获取收藏的帖子
      tags:
      - 用户管理
  /api/v1/user/{id}/fans:
    get:
      consumes:
      - multipart/form-data
      description: 获取粉丝
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取粉丝
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/utils.Error'
      summary: 获取粉丝
      tags:
      - 用户管理
  /api/v1/user/{id}/like:
    get:
      consumes:
      - multipart/form-data
      description: 获取点赞的帖子
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取点赞的帖子
          schema:
            items:
              $ref: '#/definitions/model.Post'
            type: array
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/utils.Error'
      summary: 获取点赞的帖子
      tags:
      - 用户管理
  /api/v1/user/{id}/post:
    get:
      consumes:
      - multipart/form-data
      description: 获取已发布的帖子
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取已发布的帖子
          schema:
            items:
              $ref: '#/definitions/model.Post'
            type: array
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/utils.Error'
      summary: 获取已发布的帖子
      tags:
      - 用户管理
  /api/v1/user/{id}/subscribe:
    get:
      consumes:
      - multipart/form-data
      description: 获取关注的人
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取关注的人
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/utils.Error'
      summary: 获取关注的人
      tags:
      - 用户管理
  /api/v1/user/avatar:
    put:
      consumes:
      - multipart/form-data
      description: 更新头像
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 头像文件
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 成功更新头像
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/utils.Error'
      summary: 更新头像
      tags:
      - 用户管理
  /api/v1/user/identity:
    put:
      consumes:
      - multipart/form-data
      description: 更新身份
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 身份
        in: query
        name: identity
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功更新身份
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/utils.Error'
      summary: 更新身份
      tags:
      - 用户管理
  /api/v1/user/info:
    put:
      consumes:
      - multipart/form-data
      description: 更新基本信息
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 昵称
        in: formData
        name: nickname
        required: true
        type: string
      - description: 简介
        in: formData
        name: intro
        required: true
        type: string
      - description: 年龄
        in: formData
        name: age
        required: true
        type: string
      - description: 性别
        in: formData
        name: gender
        required: true
        type: string
      - description: 身高
        in: formData
        name: stature
        required: true
        type: string
      - description: 地址
        in: formData
        name: address
        required: true
        type: string
      - description: 经验/经历
        in: formData
        name: experience
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功更新基本信息
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/utils.Error'
      summary: 更新基本信息
      tags:
      - 用户管理
  /api/v1/user/subscribe:
    delete:
      consumes:
      - multipart/form-data
      description: 取消关注用户
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 被取消关注用户ID
        in: query
        name: idolID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功取消关注用户
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/utils.Error'
      summary: 取消关注用户
      tags:
      - 用户管理
    post:
      consumes:
      - multipart/form-data
      description: 关注用户
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 被关注用户ID
        in: query
        name: idolID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功关注用户
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/utils.Error'
      summary: 关注用户
      tags:
      - 用户管理
schemes:
- http
swagger: "2.0"